Dynamic Resolution Changes
Currently, the variables stay the same; However, the scene is re-instanced every time the screen resolution changes. This system is meant for games that do not have "resizable" enabled. All resolution changes should be handled in GDScript. A UI example has been provided. A future commit will change the size of the dots and lines based upon the size of the screen, making it a percentage, rather than a fixed value.